refactors
finished the list_masters refactor, and start on making the MasterProxy
be a task.

The poll loop of the ListMastersTask is a little tricky, because it
would probably be better if the loop attempted to finish all in flight
RPCs before moving on to send new RPCs and flushing. This implies that
the Sink + Stream paradigm for Connection would be better, since it
allows the controller of the Connection to do reads and writes
independently (vs. just Sink, where reads and writes are cranked during
poll_complete).

If the MasterProxyTask is responsible for reading off a queue and
proxying to the leader connection, it's going to need to handle a
buffered queue of RPCs while the leader master is being found. This
implies that the buffer would need to / should timeout buffered RPCs who
outlive their timeout. However, we have no way of doing that for RPCs
which are stuck buffered in the mpsc queue.  Thus, callers will always
need to set their own timeout when it's appropriate.  This is the only
way time guarantees can be made when mpsc queues are in use.